Update homeWork5.js
Two text boxes were created in which the values ​​from the first and the corresponding button are passed to the second box. There are also two buttons for clearing text fields. Selector for choosing the font color in the second text box.
Fixed:
The font was changed because the text output field did not work, and the output field itself was replaced.